requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pandas==2.2.1
openpyxl==3.1.2
scrapy==2.11.1
selenium==4.18.1
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
pytest==8.0.2
pytest-django==4.8.0 